+++ test/models/design_print_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class DesignPrintTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end